+++ src/normal_facade.rs
@@ -1,0 +1,153 @@
+use crate::error::RsmqError;
+use crate::functions::RsmqFunctions;
+use crate::r#trait::RsmqConnection;
+use crate::types::{RsmqMessage, RsmqOptions, RsmqQueueAttributes};
+use core::marker::PhantomData;
+
+struct RedisConnection(redis::aio::Connection);
+
+impl std::fmt::Debug for RedisConnection {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        write!(f, "RedisAsyncConnnection")
+    }
+}
+
+#[derive(Debug)]
+pub struct Rsmq {
+    connection: RedisConnection,
+    functions: RsmqFunctions<redis::aio::Connection>,
+}
+
+impl Rsmq {
+    /// Creates a new RSMQ instance, including its connection
+    pub async fn new(options: RsmqOptions) -> Result<Rsmq, RsmqError> {
+        let password = if let Some(password) = options.password.clone() {
+            format!("redis:{}@", password)
+        } else {
+            "".to_string()
+        };
+
+        let url = format!(
+            "redis://{}{}:{}/{}",
+            password, options.host, options.port, options.db
+        );
+
+        let client = redis::Client::open(url)?;
+
+        let connection = client.get_async_connection().await?;
+
+        Ok(Rsmq::new_with_connection(options, connection))
+    }
+
+    /// Special method for when you already have a redis-rs connection and you don't want redis_async to create a new one.
+    pub fn new_with_connection(options: RsmqOptions, connection: redis::aio::Connection) -> Rsmq {
+        Rsmq {
+            connection: RedisConnection(connection),
+            functions: RsmqFunctions {
+                ns: options.ns.clone(),
+                realtime: options.realtime,
+                conn: PhantomData,
+            },
+        }
+    }
+}
+
+#[async_trait::async_trait]
+impl RsmqConnection for Rsmq {
+    async fn change_message_visibility(
+        &mut self,
+        qname: &str,
+        message_id: &str,
+        seconds_hidden: u64,
+    ) -> Result<(), RsmqError> {
+        self.functions
+            .change_message_visibility(&mut self.connection.0, qname, message_id, seconds_hidden)
+            .await
+    }
+
+    async fn create_queue(
+        &mut self,
+        qname: &str,
+        seconds_hidden: Option<u32>,
+        delay: Option<u32>,
+        maxsize: Option<i32>,
+    ) -> Result<(), RsmqError> {
+        self.functions
+            .create_queue(
+                &mut self.connection.0,
+                qname,
+                seconds_hidden,
+                delay,
+                maxsize,
+            )
+            .await
+    }
+
+    async fn delete_message(&mut self, qname: &str, id: &str) -> Result<bool, RsmqError> {
+        self.functions
+            .delete_message(&mut self.connection.0, qname, id)
+            .await
+    }
+    async fn delete_queue(&mut self, qname: &str) -> Result<(), RsmqError> {
+        self.functions
+            .delete_queue(&mut self.connection.0, qname)
+            .await
+    }
+    async fn get_queue_attributes(
+        &mut self,
+        qname: &str,
+    ) -> Result<RsmqQueueAttributes, RsmqError> {
+        self.functions
+            .get_queue_attributes(&mut self.connection.0, qname)
+            .await
+    }
+
+    async fn list_queues(&mut self) -> Result<Vec<String>, RsmqError> {
+        self.functions.list_queues(&mut self.connection.0).await
+    }
+
+    async fn pop_message(&mut self, qname: &str) -> Result<Option<RsmqMessage>, RsmqError> {
+        self.functions
+            .pop_message(&mut self.connection.0, qname)
+            .await
+    }
+
+    async fn receive_message(
+        &mut self,
+        qname: &str,
+        seconds_hidden: Option<u64>,
+    ) -> Result<Option<RsmqMessage>, RsmqError> {
+        self.functions
+            .receive_message(&mut self.connection.0, qname, seconds_hidden)
+            .await
+    }
+
+    async fn send_message(
+        &mut self,
+        qname: &str,
+        message: &str,
+        delay: Option<u64>,
+    ) -> Result<String, RsmqError> {
+        self.functions
+            .send_message(&mut self.connection.0, qname, message, delay)
+            .await
+    }
+
+    async fn set_queue_attributes(
+        &mut self,
+        qname: &str,
+        seconds_hidden: Option<u64>,
+        delay: Option<u64>,
+        maxsize: Option<i64>,
+    ) -> Result<RsmqQueueAttributes, RsmqError> {
+        self.functions
+            .set_queue_attributes(
+                &mut self.connection.0,
+                qname,
+                seconds_hidden,
+                delay,
+                maxsize,
+            )
+            .await
+    }
+}
